Clustering!

+++ hw5_release/utils.py
@@ -11,7 +11,7 @@ import os
 def visualize_mean_color_image(img, segments):
 
     img = img_as_float(img)
-    k = np.max(segments) + 1
+    k = np.max(segments.astype(np.int)) + 1
     mean_color_img = np.zeros(img.shape)
 
     for i in range(k):
